+++ test/cli.test.js
