+++ test/docker-service.test.js
